[stack_processor]: Add component config dependencies
