+++ internal/stack/stack_processor_utils.go
@@ -1,0 +1,113 @@
+package stack
+
+import (
+	u "github.com/cloudposse/terraform-provider-utils/internal/utils"
+	"os"
+	"path"
+	"path/filepath"
+	"sort"
+	"strings"
+)
+
+func findComponentStacks(componentType string,
+	component string,
+	baseComponent string,
+	componentStackMap map[string]map[string][]string) ([]string, error) {
+
+	var stacks []string
+
+	if componentStackConfig, componentStackConfigExists := componentStackMap[componentType]; componentStackConfigExists {
+		if componentStacks, componentStacksExist := componentStackConfig[component]; componentStacksExist {
+			stacks = append(stacks, componentStacks...)
+		}
+
+		if baseComponent != "" {
+			if baseComponentStacks, baseComponentStacksExist := componentStackConfig[baseComponent]; baseComponentStacksExist {
+				stacks = append(stacks, baseComponentStacks...)
+			}
+		}
+	}
+
+	unique := u.UniqueStrings(stacks)
+	sort.Strings(unique)
+	return unique, nil
+}
+
+func createComponentStackMap(filePath string) (map[string]map[string][]string, error) {
+	stackComponentMap := map[string]map[string][]string{}
+	stackComponentMap["terraform"] = map[string][]string{}
+	stackComponentMap["helmfile"] = map[string][]string{}
+
+	componentStackMap := map[string]map[string][]string{}
+	componentStackMap["terraform"] = map[string][]string{}
+	componentStackMap["helmfile"] = map[string][]string{}
+
+	dir := path.Dir(filePath)
+
+	err := filepath.Walk(dir,
+		func(p string, info os.FileInfo, err error) error {
+			if err != nil {
+				return err
+			}
+
+			isDirectory, err := u.IsDirectory(p)
+			if err != nil {
+				return err
+			}
+
+			isYaml := u.IsYaml(p)
+
+			if !isDirectory && isYaml {
+				config, _, err := ProcessYAMLConfigFile(p, &[]string{})
+				if err != nil {
+					return err
+				}
+
+				finalConfig, err := ProcessConfig(p, config, false, "", nil)
+				if err != nil {
+					return err
+				}
+
+				if componentsConfig, componentsConfigExists := finalConfig["components"]; componentsConfigExists {
+					componentsSection := componentsConfig.(map[string]interface{})
+					stackName := strings.Replace(p, dir+"/", "", 1)
+
+					if terraformConfig, terraformConfigExists := componentsSection["terraform"]; terraformConfigExists {
+						terraformSection := terraformConfig.(map[string]interface{})
+
+						for k := range terraformSection {
+							stackComponentMap["terraform"][stackName] = append(stackComponentMap["terraform"][stackName], k)
+						}
+					}
+
+					if helmfileConfig, helmfileConfigExists := componentsSection["helmfile"]; helmfileConfigExists {
+						helmfileSection := helmfileConfig.(map[string]interface{})
+
+						for k := range helmfileSection {
+							stackComponentMap["helmfile"][stackName] = append(stackComponentMap["helmfile"][stackName], k)
+						}
+					}
+				}
+			}
+
+			return nil
+		})
+
+	if err != nil {
+		return nil, err
+	}
+
+	for stack, components := range stackComponentMap["terraform"] {
+		for _, component := range components {
+			componentStackMap["terraform"][component] = append(componentStackMap["terraform"][component], strings.Replace(stack, ".yaml", "", 1))
+		}
+	}
+
+	for stack, components := range stackComponentMap["helmfile"] {
+		for _, component := range components {
+			componentStackMap["helmfile"][component] = append(componentStackMap["helmfile"][component], strings.Replace(stack, ".yaml", "", 1))
+		}
+	}
+
+	return componentStackMap, nil
+}
